requirements.txt
groq
python-dotenv
fastapi
uvicorn
pydantic
